seperate controller and modify seed.js
