+++ routes/admin.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const adminController = require('../controllers/adminController');
+const productController = require('../controllers/productController')
 const { uploadSingle, uploadMultiple } = require('../middleware/multer');
 const auth = require('../middleware/auth');
 
@@ -8,6 +9,7 @@ router.post("/signin", adminController.actionSignin);
 router.use(auth);
 router.get("/logout", adminController.actionLogout);
 
+// ketika login succes akan mengarahkan ke screen view dashboard
 router.get("/dashboard", adminController.viewDashboard);
 router.post("/dashboard/addBook", adminController.addBook);
 router.post("/dashboard/addTrans", adminController.addTrans);
@@ -18,8 +20,8 @@ router.get("/member", adminController.viewMember);
 router.get("/transaction", adminController.viewTransaction);
 
 // router.get("/barcode", adminController.viewBarcode);
-router.get("/product", adminController.viewProduct);
-router.post("/product", uploadSingle, adminController.addProduct);
-router.put("/product", uploadSingle, adminController.editProduct);
-router.delete('/product/:id', adminController.deleteProduct);
+router.get("/product", productController.viewProduct);
+router.post("/product", uploadSingle, productController.addProduct);
+router.put("/product", uploadSingle, productController.editProduct);
+router.delete('/product/:id', productController.deleteProduct);
 module.exports = router;
