+++ seed.js
@@ -37,7 +37,7 @@ var data = [
     'documents': [
       {
         _id: mongoose.Types.ObjectId('5e96cbe292b97300fc903345'),
-        name: 'admin',
+        username: 'admin',
         password: 'rahasia',
         status: 'active',
       },
