+++ views/admin/product/table_product.ejs
@@ -29,6 +29,10 @@
             <td><%= i + 1 %></td>
             <td><%= product[i].name %></td>
             <td><%= product[i].merk %></td>
+            <% for(let i = 0; i < product.length; i++) { %>
+
+            <% }%>
+
             <td><%= product[i].type %></td>
             <td><%= product[i].status %></td>
             <td><%= product[i].price %></td>
